better structure; test file; revert direct destructing from pacakage
