+++ .gitignore
@@ -1,0 +1,2 @@
+pack-lock.json
+node_modules
