+++ src/errors/KrunkerError.js
